backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
cryptography==45.0.5
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==2.8
inflection==0.5.1
iniconfig==2.1.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
pytest==7.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.22.0
requests-oauthlib==2.0.0
responses==0.22.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
toml==0.10.2
types-toml==0.10.8.20240310
uritemplate==4.2.0
urllib3==1.25.11
